moved PokemonDetail and fixed error msg bug
