+++ src/components/PokemonDetails/PokemonDetail.vue
